Remove demoQueries + cleanup (#2737)
remove demoQueries + cleanup
